pyproject.toml
[tool.poetry]
name = "xtools"
version = "0.1.1"
description = "A python package of tools by Lee Zhang"
authors = ["niamail <xggsuser@163.com>"]
readme = "README.md"
license = "LICENSE"

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
